Added decoder and refactored transformer files
